+++ src/workflows/burn.ts
@@ -1,4 +1,4 @@
-import { Signer } from '@ethersproject/abstract-signer';
+import { Signer as L1Signer } from '@ethersproject/abstract-signer';
 import {
   TransfersApi,
   CreateTransferResponseV1,
@@ -7,10 +7,16 @@ import {
 import { serializeSignature, sign, signRaw } from '../utils';
 import { BurnAddress } from './constants';
 import { GetSignableBurnRequest } from './types';
-import { StarkWallet } from '../types';
+import { StarkWallet, WalletConnection } from '../types';
 
+type burnWorkflowWithSignerRequest = WalletConnection & {
+  request: GetSignableBurnRequest;
+  transfersApi: TransfersApi;
+};
+
+/** @deprecated */
 export async function burnWorkflow(
-  signer: Signer,
+  signer: L1Signer,
   starkWallet: StarkWallet,
   request: GetSignableBurnRequest,
   transfersApi: TransfersApi,
@@ -38,6 +44,62 @@ export async function burnWorkflow(
 
   // Obtain Ethereum Address from signer
   const ethAddress = (await signer.getAddress());
+
+  // Assemble transfer params
+  const transferSigningParams = {
+    sender_stark_key: signableResult.data.sender_stark_key!,
+    sender_vault_id: signableResult.data.sender_vault_id,
+    receiver_stark_key: signableResult.data.receiver_stark_key,
+    receiver_vault_id: signableResult.data.receiver_vault_id,
+    asset_id: signableResult.data.asset_id,
+    amount: signableResult.data.amount,
+    nonce: signableResult.data.nonce,
+    expiration_timestamp: signableResult.data.expiration_timestamp,
+    stark_signature: starkSignature,
+  };
+
+  // create transfer
+  const response = await transfersApi.createTransferV1({
+    createTransferRequest: transferSigningParams,
+    xImxEthAddress: ethAddress,
+    xImxEthSignature: ethSignature,
+  });
+
+  return {
+    sent_signature: response?.data.sent_signature,
+    status: response?.data.status?.toString(),
+    time: response?.data.time,
+    transfer_id: response?.data.transfer_id,
+  };
+}
+
+export async function burnWorkflowWithSigner({
+  l1Signer,
+  l2Signer,
+  request,
+  transfersApi,
+}: burnWorkflowWithSignerRequest): Promise<CreateTransferResponseV1> {
+  // Get signable response for transfer
+  const signableResult = await transfersApi.getSignableTransferV1({
+    getSignableTransferRequest: {
+      sender: request.sender,
+      token: request.token,
+      amount: request.amount,
+      receiver: BurnAddress.BurnEthAddress,
+    },
+  });
+
+  const { signable_message: signableMessage, payload_hash: payloadHash } =
+    signableResult.data;
+
+  // Sign message with L1 credentials
+  const ethSignature = await signRaw(signableMessage, l1Signer);
+
+  // Sign hash with L2 credentials
+  const starkSignature = await l2Signer.signMessage(payloadHash);
+
+  // Obtain Ethereum Address from signer
+  const ethAddress = (await l1Signer.getAddress());
 
   // Assemble transfer params
   const transferSigningParams = {