+++ src/workflows/workflows.ts
@@ -26,7 +26,11 @@ import {
   depositERC721Workflow,
   depositEthWorkflow,
 } from './deposit';
-import { burnWorkflow, getBurnWorkflow } from './burn';
+import {
+  burnWorkflow,
+  getBurnWorkflow,
+  burnWorkflowWithSigner,
+} from './burn';
 import {
   completeERC20WithdrawalWorfklow,
   completeERC721WithdrawalWorkflow,
@@ -91,7 +95,6 @@ export class Workflows {
     return registerOffchainWorkflow(signer, starkWallet, this.usersApi);
   }
 
-
   public registerOffchainWithSigner(l1Signer: L1Signer, l2Signer: L2Signer) {
     return registerOffchainWorkflowWithSigner(l1Signer, l2Signer, this.usersApi);
   }
@@ -134,12 +137,17 @@ export class Workflows {
     );
   }
 
+  /** @deprecated */
   public burn(
     signer: Signer,
     starkWallet: StarkWallet,
     request: UnsignedBurnRequest,
   ) {
     return burnWorkflow(signer, starkWallet, request, this.transfersApi);
+  }
+
+  public burnWithSigner(walletConnection: WalletConnection, request: UnsignedBurnRequest) {
+    return burnWorkflowWithSigner({ ...walletConnection, request, transfersApi: this.transfersApi });
   }
 
   public getBurn(request: TransfersApiGetTransferRequest) {